modifying end_at column in card_mandate table (#25696)

+++ database/migrations/2021_01_05_070902_create_card_mandates_table.php
@@ -70,7 +70,7 @@ class CreateCardMandatesTable extends Migration
             $table->unsignedInteger(Entity::START_AT)
                   ->nullable();
 
-            $table->unsignedInteger(Entity::END_AT)
+            $table->bigInteger(Entity::END_AT)
                   ->nullable();
 
             $table->Integer(Entity::TOTAL_CYCLES)
